reduce speed factor to 3 use 25 as default speed

+++ web-bundle/src/main/java/com/graphhopper/helpers/DataUpdater.java
@@ -200,7 +200,7 @@ public class DataUpdater {
 
                 edgeId = qr.getClosestEdge().getEdge();
             } else {
-                edgeId = Integer.parseInt(entry.getEdgeId());
+                edgeId = (int)Double.parseDouble(entry.getEdgeId()) ;
             }
 
             if (edgeIds.contains(edgeId)) {
